packages and shipments with sales order improvements

+++ src/main/java/com/erpflow/dao/SalesOrderItemDAO.java
@@ -5,7 +5,11 @@ import com.erpflow.model.SalesOrder;
 import com.erpflow.model.SalesOrderItem;
 import com.erpflow.util.DBConnection;
 
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,7 +19,12 @@ public class SalesOrderItemDAO {
 
         String sql = """
                 INSERT INTO sales_order_items
-                (quantity, sellingPrice, item_id, sales_order_id)
+                (
+                    quantity,
+                    sellingPrice,
+                    item_id,
+                    sales_order_id
+                )
                 VALUES (?, ?, ?, ?)
                 """;
 
@@ -38,7 +47,7 @@ public class SalesOrderItemDAO {
                     salesOrderItem.getSellingPrice()
             );
 
-            statement.setLong(
+            statement.setInt(
                     3,
                     salesOrderItem.getItem().getId()
             );
@@ -53,6 +62,7 @@ public class SalesOrderItemDAO {
             try (ResultSet rs = statement.getGeneratedKeys()) {
 
                 if (rs.next()) {
+
                     salesOrderItem.setId(
                             rs.getInt(1)
                     );
@@ -60,6 +70,7 @@ public class SalesOrderItemDAO {
             }
 
         } catch (SQLException e) {
+
             throw new RuntimeException(
                     "Error saving sales order item",
                     e
@@ -67,8 +78,10 @@ public class SalesOrderItemDAO {
         }
     }
 
+
     public List<SalesOrderItem> findBySalesOrder(
-            SalesOrder salesOrder) {
+            SalesOrder salesOrder
+    ) {
 
         String sql = """
                 SELECT
@@ -77,6 +90,7 @@ public class SalesOrderItemDAO {
                     soi.sellingPrice,
                     soi.item_id,
                     soi.sales_order_id,
+
                     i.name,
                     i.description,
                     i.purchase_price,
@@ -84,13 +98,17 @@ public class SalesOrderItemDAO {
                     i.selling_price,
                     i.sku,
                     i.status
+
                 FROM sales_order_items soi
+
                 JOIN items i
                     ON soi.item_id = i.item_id
+
                 WHERE soi.sales_order_id = ?
                 """;
 
-        List<SalesOrderItem> items = new ArrayList<>();
+        List<SalesOrderItem> items =
+                new ArrayList<>();
 
         try (
                 Connection connection = DBConnection.getConnection();
@@ -119,10 +137,14 @@ public class SalesOrderItemDAO {
                     );
 
                     orderItem.setSellingPrice(
-                            rs.getBigDecimal("sellingPrice")
+                            rs.getBigDecimal(
+                                    "sellingPrice"
+                            )
                     );
 
-                    Item item = new Item();
+
+                    Item item =
+                            new Item();
 
                     item.setId(
                             rs.getInt("item_id")
@@ -137,15 +159,21 @@ public class SalesOrderItemDAO {
                     );
 
                     item.setPurchasePrice(
-                            rs.getBigDecimal("purchase_price")
+                            rs.getBigDecimal(
+                                    "purchase_price"
+                            )
                     );
 
                     item.setReorderLevel(
-                            rs.getInt("reorder_level")
+                            rs.getInt(
+                                    "reorder_level"
+                            )
                     );
 
                     item.setSellingPrice(
-                            rs.getBigDecimal("selling_price")
+                            rs.getBigDecimal(
+                                    "selling_price"
+                            )
                     );
 
                     item.setSku(
@@ -156,14 +184,19 @@ public class SalesOrderItemDAO {
                             rs.getString("status")
                     );
 
+
                     orderItem.setItem(item);
-                    orderItem.setSalesOrder(salesOrder);
+
+                    orderItem.setSalesOrder(
+                            salesOrder
+                    );
 
                     items.add(orderItem);
                 }
             }
 
         } catch (SQLException e) {
+
             throw new RuntimeException(
                     "Error fetching sales order items",
                     e
